Move misc tree twiddling code to own module
