+++ sc2ts/__init__.py
@@ -7,3 +7,4 @@ from .metadata import *
 from .inference import *
 from .validation import *
 from .info import *
+from .tree_ops import *
